Separate out the Schedule data structure
